+++ js/services/auth-interface.js
